// gestao-integrada-rh-source/src/types.ts
export type PerfilUsuario = 'CANDIDATO' | 'FUNCIONARIO' | 'RH_ADMIN';

export type StatusAdmissao = 'PENDENTE' | 'EM_ANALISE' | 'APROVADO' | 'REJEITADO';
// ...
export interface Usuario {
  id: string;
  empresaId: string;
  nome: string;
  email: string;
  senhaHash?: string;
  perfil: PerfilUsuario;
  ativo: boolean;
  cargo?: string;
  departamento?: string;
  salarioBase?: number;
  dataAdmissao?: string;
  fotoUrl?: string;
  criadoEm: string;
  atualizadoEm: string;
}

export interface DadosCadastrais {
  id: string;
  empresaId: string;
  usuarioId: string;
  cpf: string;
  pisPasep?: string;
  dataNascimento: string;
  estadoCivil: string;
  nomeMae: string;
  telefone: string;
  
  // Endereço
  cep: string;
  logradouro: string;
  numero: string;
  complemento?: string;
  bairro: string;
  cidade: string;
  uf: string;

  // Dados Bancários
  banco: string;
  tipoConta: string;
  agencia: string;
  conta: string;
  chavePix?: string;

  // LGPD & Status
  lgpdAceito: boolean;
  lgpdAceitoEm?: string;
  lgpdIpHash?: string;
  statusAdmissao: StatusAdmissao;
  observacaoRh?: string;

  criadoEm: string;
  atualizadoEm: string;
}

export interface DocumentoAdmissao {
  id: string;
  empresaId: string;
  usuarioId: string;
  tipoDocumento: 'RG_CNH' | 'COMPROVANTE_RESIDENCIA' | 'CTPS' | 'FOTO_3X4' | 'CERTIDAO_DEPENDENTE' | 'OUTROS';
  urlArquivo: string;
  nomeArquivo: string;
  tamanhoKb?: number;
  status: 'PENDENTE' | 'EM_ANALISE' | 'APROVADO' | 'REJEITADO';
  observacaoRh?: string;
  criadoEm: string;
}